Migrate to GitHub Actions
This replaces the previous Travis CI (no longer active) with GitHub Actions.

It adds a bunch of more recent Rubies (3.0, 3.1, JRuby 9.3) to the CI matrix.

It also addresses an issue in the specs due to a bundler change, by allowing the output to match a quoted or unquoted group that is being excluded.

Finally, it addresses a very noisy deprecation error.